+++ inc/zfparray1.h
@@ -23,7 +23,7 @@ public:
     blksize(block_size(rate)),
     bytes(0),
     data(0),
-    codec(stream, 0, CHAR_BIT * blksize),
+    codec(stream, 0, uint(CHAR_BIT * blksize)),
     cache(lines(csize, n)),
     dims(0)
   {
@@ -101,7 +101,7 @@ public:
   {
     off_t offset = 0;
     const uchar* d = dims;
-    for (uint i = 0; i < m; i++, p += 4, offset += blksize) {
+    for (uint i = 0; i < m; i++, p += 4, offset += off_t(blksize)) {
       uint b = block(i);
       const CacheLine* line = cache.lookup(b + 1);
       if (line)
@@ -118,7 +118,7 @@ public:
   {
     off_t offset = 0;
     const uchar* d = dims;
-    for (uint i = 0; i < m; i++, p += 4, offset += blksize) {
+    for (uint i = 0; i < m; i++, p += 4, offset += off_t(blksize)) {
       stream.seek(offset);
       codec.encode(p, 1, d ? *d++ : 0);
       stream.flush();
@@ -223,12 +223,12 @@ protected:
     if (c != b) {
       if (t.dirty()) {
         // write back dirty cache line
-        stream.seek(c * blksize);
+        stream.seek(off_t(c * blksize));
         codec.encode(p->a, 1, dims ? dims[c] : 0);
         stream.flush();
       }
       // fetch cache line
-      stream.seek(b * blksize);
+      stream.seek(off_t(b * blksize));
       codec.decode(p->a, 1);
     }
     return p;
@@ -266,7 +266,7 @@ protected:
   // number of cache lines corresponding to size (or suggested size if zero)
   static uint lines(size_t size, uint n)
   {
-    n = (size ? size : 8 * sizeof(Scalar)) / sizeof(CacheLine);
+    n = uint((size ? size : 8 * sizeof(Scalar)) / sizeof(CacheLine));
     return std::max(n, 1u);
   }
